requirements.txt
python-dotenv
discord.py
firebase_admin
requests
sentry-sdk